requirements.txt
asgiref==3.10.0
Django==5.2.7
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
six==1.17.0
sqlparse==0.5.3
